+++ packages/Chem/src/package-test.ts
@@ -1,5 +1,5 @@
 import * as DG from 'datagrok-api/dg';
-import {runTests, TestContext, tests} from '@datagrok-libraries/utils/src/test';
+import {runTests, TestContext, tests, initAutoTests as initTests } from '@datagrok-libraries/utils/src/test';
 
 import './tests/ui-tests-info-panel';
 import './tests/ui-tests-top-menu';
@@ -48,4 +48,9 @@ export {tests};
 export async function test(category: string, test: string, testContext: TestContext, stressTest?: boolean): Promise<DG.DataFrame> {
   const data = await runTests({category, test, testContext, stressTest});
   return DG.DataFrame.fromObjects(data)!;
+}
+
+//name: initAutoTests
+export async function initAutoTests() {
+  await initTests(_package, _package.getModule('package-test.js'));
 }