+++ packages/CruddyDemo/src/package-test.ts
@@ -1,5 +1,5 @@
 import * as DG from "datagrok-api/dg";
-import {runTests, tests, TestContext} from '@datagrok-libraries/utils/src/test';
+import {runTests, tests, TestContext, initAutoTests as initTests } from '@datagrok-libraries/utils/src/test';
 
 export let _package = new DG.Package();
 export {tests};
@@ -14,4 +14,9 @@ import './cruddy-test';
 export async function test(category: string, test: string, testContext: TestContext): Promise<DG.DataFrame> {
   const data = await runTests({category, test, testContext});
   return DG.DataFrame.fromObjects(data)!;
+}
+
+//name: initAutoTests
+export async function initAutoTests() {
+  await initTests(_package, _package.getModule('package-test.js'));
 }