+++ tools/bin/utils/utils.ts
@@ -1,5 +1,9 @@
 import fs from 'fs';
 import path from 'path';
+import { exec } from 'child_process';
+import { promisify } from 'util';
+
+const execAsync = promisify(exec);
 
 
 /* Waits [ms] milliseconds */
@@ -32,7 +36,7 @@ export function wordsToCamelCase(s: string, firstUpper: boolean = true): string 
 }
 
 export function camelCaseToKebab(s: string): string {
-  return s.replace(/[A-Z]/g, (char: string, index: number) => index == 0 ? char.toLowerCase() : '-'+ char.toLowerCase());
+  return s.replace(/[A-Z]/g, (char: string, index: number) => index == 0 ? char.toLowerCase() : '-' + char.toLowerCase());
 }
 
 export function removeScope(name: string): string {
@@ -42,9 +46,9 @@ export function removeScope(name: string): string {
 
 export function mapURL(conf: Config): Indexable {
   const urls: Indexable = {};
-  for (const server in conf['servers']) 
+  for (const server in conf['servers'])
     urls[conf['servers'][server]['url']] = server;
-  
+
   return urls;
 }
 
@@ -134,9 +138,9 @@ export const queryExtension = '.sql';
 export const scriptExtensions = ['.jl', '.m', '.py', '.R'];
 export function checkScriptLocation(filepath: string): boolean {
   if (!(filepath.startsWith('scripts/') || filepath.startsWith('projects/') || filepath.startsWith('dockerfiles/')) &&
-    scriptExtensions.some((ext: any) => filepath.endsWith(ext))) 
+    scriptExtensions.some((ext: any) => filepath.endsWith(ext)))
     return false;
-  
+
   return true;
 };
 
@@ -197,7 +201,7 @@ export function getScriptInputs(script: string, comment: string = '#'): object[]
   for (const match of script.matchAll(regex)) {
     const type = dgToTsTypeMap[match[1]] || 'any';
     const name = match[2];
-    inputs.push({type, name});
+    inputs.push({ type, name });
   }
   return inputs;
 };
@@ -225,3 +229,35 @@ export interface Config {
 }
 
 export interface Indexable { [key: string]: any }
+
+
+
+export async function runScript(script: string, path: string, verbose: boolean = false) {
+  try {
+    const { stdout, stderr } = await execAsync(script, { cwd: path });
+    if (stderr && verbose) {
+      console.error(`Warning/Error: ${stderr}`);
+    }
+    if (stdout && verbose) {
+      console.log(`Output: ${stdout}`);
+    }
+  } catch (error: any) {
+    console.error(`Execution failed: ${error.message}`);
+    throw new Error("Cant run script");
+  }
+}
+
+export function setHost(host: any, configFile: any) {
+  if (host) {
+    if (host in configFile.servers) {
+      process.env.HOST = host;
+      console.log('Environment variable `HOST` is set to', host);
+    } else {
+      console.error(`Unknown server alias. Please add it to Config File`);
+      return false;
+    }
+  } else if (configFile.default) {
+    process.env.HOST = configFile.default;
+    console.log('Environment variable `HOST` is set to', configFile.default);
+  }
+}